# backend/docs/openapi/api.yaml
...
openapi: 3.0.3
info:
    description: |
        Machine-generated OpenAPI specification for AppOS APIs, with ext routes generated from route code using group-matrix as single source of truth.
    title: AppOS API
    version: 0.1.0
...
components:
    schemas:
        ErrorEnvelope:
            properties:
                code:
                    type: integer
                message:
                    type: string
            required:
                - message
            type: object
        GenericRequest:
            additionalProperties: true
            description: Generic request payload placeholder (refine per endpoint)
            type: object
...
        SuccessEnvelope:
            additionalProperties: true
            description: Generic success payload envelope
            example:
                ok: true
            type: object
    securitySchemes:
        bearerAuth:
            bearerFormat: PocketBase token
            scheme: bearer
            type: http
paths:
...
    /api/ext/tunnel/servers/{id}/setup:
        get:
            operationId: get_api_ext_tunnel_servers_id_setup
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
            security:
                - bearerAuth: []
            summary: Get tunnel servers by id setup
            tags:
                - Tunnel
    /api/ext/tunnel/servers/{id}/status:
        get:
            operationId: get_api_ext_tunnel_servers_id_status
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
            security:
                - bearerAuth: []
            summary: Get tunnel servers by id status
            tags:
                - Tunnel
    /api/ext/tunnel/servers/{id}/token:
        post:
            operationId: post_api_ext_tunnel_servers_id_token
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GenericRequest'
                required: false
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
            security:
                - bearerAuth: []
            summary: Create or execute tunnel servers by id token
            tags:
                - Tunnel
...